requirements.txt
discord~=2.1.0
google_api_python_client==2.74.0
google_auth_oauthlib==0.8.0
protobuf==4.21.12
